requirements.txt
appdirs==1.4.3
APScheduler==3.6.3
attrs==19.3.0
beautifulsoup4==4.8.2
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
idna==2.8
importlib-metadata==1.3.0
more-itertools==8.0.2
packaging==19.2
pathspec==0.6.0
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.6
pytest==5.3.2
python-dotenv==0.10.3
pytz==2019.3
regex==2019.12.20
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.12
toml==0.10.0
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
